requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
google-generativeai==0.8.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3